Capstone_design2/backend/requirements/common.txt
django~=3.2.18
djangorestframework
djangorestframework-jwt
django-cors-headers
pillow
torch~=1.13.1
numpy